requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
apex=0.1=pypi_0
cudatoolkit=10.2.89=hfd86e86_1
datasets=1.11.0=pypi_0
numpy=1.20.3=py38hf144106_0
numpy-base=1.20.3=py38h74d4b33_0
python=3.8.11=h12debd9_0_cpython
pytorch=1.8.1=py3.8_cuda10.2_cudnn7.6.5_0
pytorch-ignite=0.4.6=pypi_0
scikit-learn=0.24.2=py38ha9443f7_0
tokenizers=0.10.3=pypi_0
transformers=4.8.2=pypi_0
